Created new directory events file manager extension, added label to symlinks in navigator, bugfixes
